backend/requirements.txt
django==3.2
djongo==1.3.6
sqlparse==0.2.4
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
pymongo==3.12.3
pytz>=2021.1
djangorestframework-simplejwt
